Created using Colab
In this assignment, we explored using Machine Learning (ML) services to perform sentiment analysis on a dataset of IMDb movie reviews. The objective was to classify movie reviews as positive or negative. Initially, we examined the feasibility of using an existing Natural Language Processing (NLP) service, such as the Google Cloud Natural Language API, and then compared its performance with a custom-built model using TensorFlow.